third_party/blink/web_tests/external/wpt/css/css-overflow/column-scroll-marker-focus-004-expected.txt

This is a testharness.js-based test.
[FAIL] Focus column #1
  assert_equals: expected Element node <div tabindex="0" id="block2">\n      block<br>\n      bloc... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #10
  assert_equals: expected Element node <div id="after" tabindex="0">after</div> but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #2
  assert_equals: expected Element node <div tabindex="0" id="inlineBlock1" style="display:inline... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #9
  assert_equals: expected Element node <div id="after" tabindex="0">after</div> but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #8
  assert_equals: expected Element node <span tabindex="0" id="inlineElm2">\n      inline<br>\n    ... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #7
  assert_equals: expected Element node <div tabindex="0" id="inlineBlock2" style="display:inline... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #6
  assert_equals: expected Element node <div tabindex="0" id="inlineBlock2" style="display:inline... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #5
  assert_equals: expected Element node <div tabindex="0" id="inlineBlock2" style="display:inline... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #4
  assert_equals: expected Element node <span tabindex="0" id="inlineElm1">\n      inline<br>\n    ... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
[FAIL] Focus column #3
  assert_equals: expected Element node <div tabindex="0" id="inlineBlock1" style="display:inline... but got Element node <div tabindex="0" id="block1" style="height:85px; backgro...
Harness: the test ran to completion.
